feat: move BrokerParams under backtesting module
