+++ src/examples/single_asset/two_ema_cross/backtest.py
@@ -1,7 +1,7 @@
 import os
 from dotenv import dotenv_values
 
-from ....strategy_tester.broker_params import BrokerParamsBuilder
+from ....strategy_tester.backtesting.broker_params import BrokerParamsBuilder
 from .pipeline import create_pipeline_backtest
 
 # to execute the script past this code in the terminal:
