+++ src/examples/single_asset/two_ema_cross/optimization.py
@@ -2,7 +2,7 @@ import os
 from dotenv import dotenv_values
 
 from ....strategy_tester.optimization_params import OptimizationParamsBuilder
-from ....strategy_tester.broker_params import BrokerParamsBuilder
+from ....strategy_tester.backtesting.broker_params import BrokerParamsBuilder
 from .pipeline import create_pipeline_optimization
 
 # to execute the script past this code in the terminal:
